Opti: fetch metrics only for shown columns.
Since user can choose what columns to show,
we can skip fetching what is not used!

For example, if the swap column is not shown,
we can skip fetching swap.

And if the App pannel is not open,
there is no need to fetch the metrics that it shows (for now).